refactor: metrics initialization changes
Instead of initializing the metrics structure and working with it directly, I decided to create the metrics interface with two main implementations.
First one is realMetrics that would be working only if metrics enabled and gathers metrics information.
And another one is noMetrics, which is dummy implemenation and does nothing serious.
The refactoring is being made with the goal of removing checking for nil variable constantly.